[MWPW-169785] Workflow authoring enhancements to improve performance (#320)
### Description:
This PR removes the need for the icon expression used for the Unity
workflow verb. Now, we’ll get the verb directly from the verb-widget
block.
This has a dependency on [this DC
PR](https://github.com/adobecom/dc/pull/1071). The DC PR needs to be
merged first so this can be tested easily.

Resolves: [MWPW-169785](https://jira.corp.adobe.com/browse/MWPW-169785)

**Test URLs:**
- Before:
https://stage--dc--adobecom.hlx.page/drafts/rbogos/acrobat/compress-pdf-new?unitylibs=stage&martech=off
- After:
https://stage--dc--adobecom.hlx.page/drafts/rbogos/acrobat/compress-pdf-new?unitylibs=mwpw-169785-performance&martech=off