[MWPW-171722]Code to handle different text for desktop in transition screen (#364)
* Code to handle different text for desktop in transition screen

Resolves: [MWPW-171722](https://jira.corp.adobe.com/browse/MWPW-171722)

**Test URLs:**
Test pages to make sure nothing breaks on live experience
- Before:
https://stage--cc--adobecom.hlx.page/products/photoshop/online?martech=off&unitylibs=stage
- After:
https://stage--cc--adobecom.hlx.page/products/photoshop/online?martech=off&unitylibs=MWPW-171722

Test page to make sure if 2 different text is authored on transition
screen fragment before CC phase code goes live doesn't break anything on
live experience:
- Before:
https://stage--cc--adobecom.hlx.page/products/photoshop/online?martech=off&unitylibs=stage
- After:
https://stage--cc--adobecom.hlx.page/products/photoshop/online?martech=off&unitylibs=splash-test

splash-test feature branch points to a transition fragment which has the
updated authoring with 2 different text. As part of this PR we need to
test and make sure that there is no change in existing experience in
both the scenarios when 2 different texts are authored or when only 1
test is authored in transition screen fragment.